fix: move tests to MainTest.java
