+++ src/test/java/MainTest.java
@@ -1,39 +1,37 @@
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
-
-import org.junit.After;
-import org.junit.Test;
-
-import java.io.BufferedReader;
-import java.io.InputStreamReader;
 import java.io.ByteArrayOutputStream;
+import java.io.File;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.io.PrintStream;
 import java.net.HttpURLConnection;
 import java.net.URL;
-import java.nio.ByteBuffer;
-import java.nio.charset.Charset;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import org.eclipse.jetty.client.api.ContentProvider;
 import org.eclipse.jetty.server.Request;
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.server.ServerConnector;
 import org.eclipse.jetty.server.handler.AbstractHandler;
+import org.junit.After;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertTrue;
+import org.junit.Test;
 
-import se.ciserver.TestUtils;
 import se.ciserver.ContinuousIntegrationServer;
 import se.ciserver.TestRunner;
-import se.ciserver.github.PushParser;
-import se.ciserver.github.Push;
+import se.ciserver.TestUtils;
+import se.ciserver.buildlist.Build;
+import se.ciserver.buildlist.BuildStore;
 import se.ciserver.github.InvalidPayloadException;
+import se.ciserver.github.Push;
+import se.ciserver.github.PushParser;
 
 /**
  * Test class
@@ -262,6 +260,100 @@ public class MainTest
 
         assertEquals("Set Commit Status failed, post request exception\n", systemOutCatcher.toString());
         System.setOut(originalOut);
+    }
+
+    private static final String TEST_FILE = "build-history-test.json";
+    
+    /**
+     * Test if build object can be created normally
+     */
+    @Test
+    public void newBuild_setsAllFields() {
+        String commitId = "abc123";
+        String branch = "assessment";
+        Boolean status = true;
+        String log = "build output";
+
+        Build b = Build.newBuild(commitId, branch, status, log);
+
+        assertNotNull("id should be generated", b.id);
+        assertFalse("id should not be empty", b.id.isEmpty());
+        assertEquals("commit id should match", commitId, b.commitId);
+        assertEquals("branch should match", branch, b.branch);
+        assertEquals("status should match", status, b.status);
+        assertEquals("log should match", log, b.log);
+        assertNotNull("timestamp should be set", b.timestamp);
+    }
+
+    /**
+     * Verify the behavior when history file does not exist yet
+     */
+    @Test
+    public void newStoreWithoutFileStartsEmpty() {
+        File f = new File(TEST_FILE);
+        if (f.exists()) {
+            assertTrue(f.delete());
+        }
+
+        BuildStore store = new BuildStore(TEST_FILE);
+
+        List<Build> all = store.getAll();
+        assertNotNull(all);
+        assertTrue("store should start empty when file is missing", all.isEmpty());
+    }
+
+    /**
+     * Checks whether the file persist and being reloaded after server restart,
+     * simulated by having another BuildStore object
+     */
+    @Test
+    public void addPersistsBuildAndCanBeReloaded() {
+        // Cleanup
+        File f = new File(TEST_FILE);
+        if (f.exists()) {
+            assertTrue(f.delete());
+        }
+
+        // first store: add one build
+        BuildStore store1 = new BuildStore(TEST_FILE);
+        Build build = Build.newBuild("commit1", "assessment",
+                                     true, "log1");
+        store1.add(build);
+
+        List<Build> all1 = store1.getAll();
+        assertEquals(1, all1.size());
+        assertEquals(build.id, all1.get(0).id);
+
+        // second store: simulate server restart by creating a new instance
+        BuildStore store2 = new BuildStore(TEST_FILE);
+        List<Build> all2 = store2.getAll();
+        assertEquals("should load one build from file", 1, all2.size());
+        Build loaded = all2.get(0);
+        assertEquals(build.id, loaded.id);
+    }
+
+    /**
+     * Test function returns the correct build with id search
+     */
+    @Test
+    public void getByIdReturnsCorrectBuildOrNull() {
+        BuildStore store = new BuildStore(TEST_FILE);
+        Build b1 = Build.newBuild("c1", "assessment", true, "log1");
+        Build b2 = Build.newBuild("c2", "assessment", false, "log2");
+
+        store.add(b1);
+        store.add(b2);
+
+        Build found1 = store.getById(b1.id);
+        assertNotNull(found1);
+        assertEquals(b1.id, found1.id);
+
+        Build found2 = store.getById(b2.id);
+        assertNotNull(found2);
+        assertEquals(b2.id, found2.id);
+
+        Build missing = store.getById("does-not-exist");
+        assertNull("unknown id should return null", missing);
     }
 
 }
